# 8259278: Optimize Vector API slice and unslice operations